+++ packages/headless/src/controllers/insight/facets/range-facet/numeric-facet/headless-insight-numeric-facet.ts
@@ -1,6 +1,6 @@
 import {NumericRangeRequest} from '../../../../../features/facets/range-facets/numeric-facet-set/interfaces/request';
 import {NumericFacetValue} from '../../../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
-import {isRangeFacetValueSelected} from '../../../../../features/facets/range-facets/generic/range-facet-utils';
+import {getInsightAnalyticsActionForToggleRangeFacetSelect} from '../../../../../features/facets/range-facets/generic/range-facet-insight-utils';
 import {
   buildCoreNumericFacet,
   buildNumericRange,
@@ -15,12 +15,8 @@ import {InsightEngine} from '../../../../../app/insight-engine/insight-engine';
 import {executeSearch} from '../../../../../features/insight-search/insight-search-actions';
 import {
   logFacetClearAll,
-  logFacetDeselect,
-  logFacetSelect,
   logFacetUpdateSort,
 } from '../../../../../features/facets/facet-set/facet-set-insight-analytics-actions';
-import {RangeFacetValue} from '../../../../../features/facets/range-facets/generic/interfaces/range-facet';
-import {FacetSelectionChangeMetadata} from '../../../../../features/facets/facet-set/facet-set-analytics-actions-utils';
 
 export type {
   NumericRangeOptions,
@@ -69,7 +65,10 @@ export function buildNumericFacet(
       coreController.toggleSelect(selection);
       dispatch(
         executeSearch(
-          getAnalyticsActionForToggleRangeFacetSelect(getFacetId(), selection)
+          getInsightAnalyticsActionForToggleRangeFacetSelect(
+            getFacetId(),
+            selection
+          )
         )
       );
     },
@@ -80,16 +79,4 @@ export function buildNumericFacet(
       };
     },
   };
-}
-
-function getAnalyticsActionForToggleRangeFacetSelect(
-  facetId: string,
-  selection: RangeFacetValue
-) {
-  const facetValue = `${selection.start}..${selection.end}`;
-  const payload: FacetSelectionChangeMetadata = {facetId, facetValue};
-
-  return isRangeFacetValueSelected(selection)
-    ? logFacetDeselect(payload)
-    : logFacetSelect(payload);
 }