+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.103.6](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.103.5...@coveo/headless@1.103.6) (2022-09-16)
+
+
+### Bug Fixes
+
+* **headless:** insightFacet and insightDateFacet controllers set to use the correct analytics actions ([#2434](https://github.com/coveo/ui-kit/issues/2434)) ([6234ace](https://github.com/coveo/ui-kit/commit/6234acee8fba53fff75868459fce17004640e54c))
+
+
+
+
+
+## [1.103.5](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.103.4...@coveo/headless@1.103.5) (2022-09-15)
+
+
+### Bug Fixes
+
+* **headless:** added smart snippet inline link analytics ([#2416](https://github.com/coveo/ui-kit/issues/2416)) ([99e5369](https://github.com/coveo/ui-kit/commit/99e5369032c0c35f0e5b57d373afe3c0ceda3e0e))
+
+
+
+
+
+## [1.103.4](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.103.3...@coveo/headless@1.103.4) (2022-09-14)
+
+
+### Bug Fixes
+
+* **headless:** log search box submit + custom event on query trigger modification ([#2404](https://github.com/coveo/ui-kit/issues/2404)) ([35dd953](https://github.com/coveo/ui-kit/commit/35dd953735f0f9d5f7c9e7f2c801d4eeaee7d98b))
+
+
+
+
+
 ## [1.103.3](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.103.2...@coveo/headless@1.103.3) (2022-09-09)